requirements.txt
lxml==5.3.0
pillow==11.0.0
python-docx==1.1.2
typing_extensions==4.12.2
